requirements/base.txt
Django==2.2.19
django-cors-headers==3.1.1
djangorestframework==3.11.2
PyJWT==1.7.1 # TODO Unsure of its dependencies/dependants
djangorestframework-simplejwt==4.3.0 # Requires Django and Django rest framework
psycopg2-binary==2.8.3
git+https://github.com/joestump/python-oauth2.git
requests==2.22.0
vine==1.3.0 # celery 4.3 requires vine 1.3.0, latest version of 5.0.0 will break
celery==4.3.0
django_celery_results==1.1.2
django_celery_beat==1.5.0
sh==1.12.14
six==1.12.0 # Python 2 and 3 compatibility library, Pytest
django-csp==3.5 # CSPMiddleware Content-Security-Policy
django-computedfields==0.1.3 # Computed fields for Django

urllib3==1.25.6
text-unidecode==1.3
python-dateutil==2.8.0 # TODO Unsure of its dependants

certifi==2019.9.11 # TODO figure out specific use cases
chardet==3.0.4 # TODO figure out specific use cases
httplib2==0.19.0 # TODO figure out specific use cases
idna==2.8 # TODO figure out specific use cases
entrypoints==0.3 # TODO figure out specific use cases

sentry-sdk==0.12.1 # Logging
django-anymail==7.0.0 # Email client
chardet==3.0.4 # Encoding estimation
dataclasses==0.8 # Email sender config
setuptools==45.1.0

pylti1p3==1.8.5